requirements.txt
streamlit
openai>=1.0.0
faiss-cpu
tqdm
PyPDF2
scikit-learn
python-dotenv
sentence-transformers